socialnetwork/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cramjam==2.8.3
distlib==0.3.8
Django==3.2.25
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-filter==2.3.0
django-mongodb-engine==0.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1  # Replace djangorestframework-jwt with djangorestframework-simplejwt
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.6.1
filelock==3.13.4
frozenlist==1.4.1
idna==3.7
Markdown==3.6
mongoengine==0.28.2
multidict==6.0.5
mysqlclient==2.2.4
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.1
pycparser==2.21
PyJWT==2.6.0  # Ensure the latest version compatible with other packages
pymongo==4.7.0
python-dotenv==1.0.1
python-snappy==0.7.1
pytz==2024.1
reportlab==4.2.0
requests==2.31.0
setuptools==69.5.1
sqlparse==0.2.4
twilio==9.0.5
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.0
yarl==1.9.4
